Change syncfolder command
